Replace Response and Request with RequestContext
This removes the Response and Request types and replaces them with a
an interface, RequestContext, that is passed to each handler.

The router, handlers, and groups accept a generic type that is used to
create a new RequestContext for each request. This allows consumers to
provide their own RequestContext implementation that can be used to
store data that should be available to all handlers in the request
like the current user, database connections, etc.