Remove metal, wrap handlers by default
This removes the metal middleware stack in favor of simplifying on the
existing middleware stack. HTTP style middleware can still be easily
used by wrapping the handler in a middleware function.

This also wraps all handlers in the middleware stack by default to avoid
handling that logic in the critical path.