Separate groups and namespaces
This change introduces the concept of namespaces, which is a group but
with a prefix that is prepended to all routes registered with the
group.

This is a breaking change, but it's a much nicer API than having
to pass empty strings to the Group method and avoids the extra
complexity of introducing functional options pattern given the usage
of generics.

Additionally, it makes it nicer to talk about to people, since groups
and namespaces often have different purposes.